requirements.txt
attrs==23.2.0
boto3==1.33.13
botocore==1.33.13
certifi==2024.2.2
cffi==1.15.1
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
importlib-metadata==6.7.0
jmespath==1.0.1
outcome==1.3.0.post0
pycparser==2.21
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
s3transfer==0.8.2
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.11.1
typing-extensions==4.7.1
urllib3==1.26.18
wsproto==1.2.0
zipp==3.15.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
google-api-python-client==2.127.0
